[rfct]: field w,hをserverの値で初期化
